requirements.txt
Appium-Python-Client~=1.0.2